requirements.txt
mysql-connector-python
Flask
Flask-Cors
python-dotenv
regex
requests

# development reqquirements
isort

-e .
